# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khabar — hyperlocal citizen journalism API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type PostCategory = (typeof PostCategory)[keyof typeof PostCategory];

export const PostCategory = {
  infrastructure: "infrastructure",
  crime: "crime",
  politics: "politics",
  event: "event",
  health: "health",
  weather: "weather",
} as const;

export type VerificationStatus =
  (typeof VerificationStatus)[keyof typeof VerificationStatus];

export const VerificationStatus = {
  unverified: "unverified",
  community: "community",
  editor: "editor",
  verified: "verified",
} as const;

export interface Location {
  lat: number;
  lng: number;
  neighborhood: string;
}

export interface ReputationEvent {
  id: number;
  date: string;
  pointsChange: number;
  reason: string;
}

export interface User {
  uid: string;
  username: string;
  displayName: string;
  photoUrl?: string | null;
  currentReputationScore: number;
  state: string;
  district: string;
  locality: string;
  postCount: number;
  verifiedCount: number;
  readCount: number;
  createdAt: string;
}

export interface CurrentUser {
  uid: string;
  username: string;
  displayName: string;
  photoUrl?: string | null;
  email: string;
  phoneNumber?: string | null;
  currentReputationScore: number;
  state: string;
  district: string;
  locality: string;
  isEmailVerified: boolean;
  isPhoneVerified: boolean;
  postCount: number;
  verifiedCount: number;
  readCount: number;
  createdAt: string;
}

export type PostCurrentUserVote =
  (typeof PostCurrentUserVote)[keyof typeof PostCurrentUserVote];

export const PostCurrentUserVote = {
  up: "up",
  down: "down",
  none: "none",
} as const;

export interface Post {
  id: number;
  authorId: string;
  authorUsername: string;
  authorName: string;
  authorPhotoUrl?: string | null;
  authorReputation: number;
  headline: string;
  details: string;
  category: PostCategory;
  isBreaking: boolean;
  location: Location;
  verificationStatus: VerificationStatus;
  upvotes: number;
  downvotes: number;
  savedByCurrentUser: boolean;
  currentUserVote: PostCurrentUserVote;
  createdAt: string;
}

export interface UserProfile {
  uid: string;
  username: string;
  displayName: string;
  photoUrl?: string | null;
  currentReputationScore: number;
  state: string;
  district: string;
  locality: string;
  postCount: number;
  verifiedCount: number;
  readCount: number;
  createdAt: string;
  reputationHistory: ReputationEvent[];
  recentPosts: Post[];
}

export type VoteBodyDirection =
  (typeof VoteBodyDirection)[keyof typeof VoteBodyDirection];

export const VoteBodyDirection = {
  up: "up",
  down: "down",
  clear: "clear",
} as const;

export interface VoteBody {
  direction: VoteBodyDirection;
}

export interface VerifyBody {
  status: VerificationStatus;
}

export interface SaveBody {
  saved: boolean;
}

export interface SaveStatus {
  postId: number;
  saved: boolean;
}

export interface TrendingLocality {
  neighborhood: string;
  postCount: number;
  breakingCount: number;
}

export interface CategoryCount {
  category: PostCategory;
  count: number;
}

export interface StatsOverview {
  totalPosts: number;
  breakingNow: number;
  verifiedToday: number;
  activeReporters: number;
  categoryCounts: CategoryCount[];
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export type RequestUploadUrlResponseMetadata = {
  name: string;
  size: number;
  contentType: string;
};

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata: RequestUploadUrlResponseMetadata;
}

export type RegisterBody = {
  /**
   * @minLength 3
   * @maxLength 24
   */
  username: string;
  email: string;
  /** @minLength 8 */
  password: string;
  /**
   * @minLength 2
   * @maxLength 50
   */
  displayName: string;
  state: string;
  district: string;
  locality: string;
  phoneNumber?: string;
  photoUrl?: string;
};

export type LoginBody = {
  email: string;
  password: string;
};

export type Logout200 = {
  ok: boolean;
};

export type UpdateCurrentUserBody = {
  displayName?: string;
  state?: string;
  district?: string;
  locality?: string;
  phoneNumber?: string;
  photoUrl?: string;
};

export type CheckAvailabilityParams = {
  username?: string;
  email?: string;
};

export type CheckAvailability200 = {
  usernameAvailable?: boolean;
  emailAvailable?: boolean;
};

export type ListTopReportersParams = {
  limit?: number;
};

export type ListPostsParams = {
  category?: PostCategory;
  authorId?: string;
  savedBy?: string;
};

export type CreatePostBody = {
  headline: string;
  details: string;
  category: PostCategory;
  isBreaking: boolean;
  neighborhood?: string;
};