# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Khabar — hyperlocal citizen journalism API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 24
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                displayName:
                  type: string
                  minLength: 2
                  maxLength: 50
                state:
                  type: string
                district:
                  type: string
                locality:
                  type: string
                phoneNumber:
                  type: string
                photoUrl:
                  type: string
              required:
                - username
                - email
                - password
                - displayName
                - state
                - district
                - locality
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict (username or email already taken)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateCurrentUser
      tags: [auth]
      summary: Update the authenticated user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                state:
                  type: string
                district:
                  type: string
                locality:
                  type: string
                phoneNumber:
                  type: string
                photoUrl:
                  type: string
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
  /auth/check-availability:
    get:
      operationId: checkAvailability
      tags: [auth]
      summary: Check whether a username or email is available
      parameters:
        - name: username
          in: query
          required: false
          schema:
            type: string
        - name: email
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Availability
          content:
            application/json:
              schema:
                type: object
                properties:
                  usernameAvailable:
                    type: boolean
                  emailAvailable:
                    type: boolean
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CurrentUser:
      type: object
      properties:
        uid:
          type: string
        username:
          type: string
        displayName:
          type: string
        photoUrl:
          type: string
          nullable: true
        email:
          type: string
        phoneNumber:
          type: string
          nullable: true
        currentReputationScore:
          type: integer
        state:
          type: string
        district:
          type: string
        locality:
          type: string
        isEmailVerified:
          type: boolean
        isPhoneVerified:
          type: boolean
        postCount:
          type: integer
        verifiedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - uid
        - username
        - displayName
        - email
        - currentReputationScore
        - state
        - district
        - locality
        - isEmailVerified
        - isPhoneVerified
        - postCount
        - verifiedCount
        - createdAt
...